Factor out Info to separate module
